Adapt Size of Header, Annotation and Graph according to %
Adds a new configuration option -- embebbed.
This object sets the header and annotation percentages.

Example: embedded: {headerPct: 15, annotationPct: 25}
This would get the header to be 15% of the height, annotations to be 25%
of the height (and, accordingly, the graph takes up 65%). If left
blank, it defaults to: {headerPct: 10, annotationPct: 25}. If
isEmbedded is set to false, the embedded object is ignored.